Refactoring and support for rates persistence
- Refactored the code.
- Added DAO, Impls and Utils for persistence

+++ src/main/java/io/xrates/backend/constants/RateProvider.java
@@ -1,11 +1,15 @@
 package io.xrates.backend.constants;
 
+/*
+ * TODO: correct the Rate Provider's id format
+ * to accommodate more countries to this system. 
+ */
 public enum RateProvider {
 	DBS(1, "DBS"),
 	OCBC(2, "OCBC"),
 	REMIT(3, "REMIT GURU");
 	
-	private final int value;
+	private final long value;
 	private final String providerName;
 	
 	private RateProvider(int value, String providerName) {
@@ -13,7 +17,7 @@ public enum RateProvider {
 		this.providerName = providerName;
 	}
 	
-	public int getValue() {
+	public long getValue() {
 		return value;
 	}
